zloop: ignore EINTR from poll if nonstop is enabled
If SIGINT signal arrives while poll is blocking, regardless of attached
signal handler it will cause the poll to return with errno EINTR.
Currently, this causes zloop to return from the event loop. For nonstop
mode to work as intended, we have to ignore EINTR in this case.

Added test cases that verify the functionality when catching an actual
SIGINT signal with and without nonstop mode. Tested in Linux and
Windows.