zbeacon: Fix setsockopt IP_MULTICAST_IF parameter type
zbeacon multicast always fails at `setsockopt` for IP_MULTICAST_IF on
Linux, because Linux expects the parameter type to be either `ip_mreqn`,
`ip_mreq`, or `in_addr` structure instead of interface index. On the
other hand, Windows does not support those structures, but instead
supports a DWORD of IPv4 address or an interface index. OSX seems to
support `in_addr`. Modified the parameter type to be interface index
for Windows and `in_addr` for others.

Multicast beacons were also received from all multicast groups instead
of just the one we join to. This is because the code bound the listening
UDP socket to `INADDR_ANY` when using multicast. Modified to bind the
listening UDP socket only to the specified multicast address.