zsock: Check native socket in zsock_resolve before ZMQ socket
Especially in Windows (at least on builds using MSVC), `zsock_resolve`
randomly (but quite often) segfaults if the passed pointer is actually
pointing to a native `SOCKET` instead of zactor, zsock, or a ZMQ socket.
This happens because `zmq_getsockopt` tries to check if, after casting,
the pointer points to an object with specific tag (`check_tag()`).
Unfortunately, reading this area of memory is often an access violation
if the pointer actually pointed to a `SOCKET` (which is a smaller data
type).

It seems that checking if the pointer is to a native socket first before
checking if it is a ZMQ socket fixes this, so these checks were
reordered.

Resolves #2300