ztimerset: Make selftest more reliable
On Windows, the `Sleep()` function resolution (which is used by
`zclock_sleep()`) can typically be as bad as 16 ms. This caused the
selftest to randomly fail because it waited exactly the remaining time
to the timer expiration before asserting the result. Made this more
reliable by adding some margin to the sleep.